Implements V2 of BadiDate library
