+++ .eslintignore
@@ -1,0 +1,3 @@
+extern/
+node_modules/
+dist/
