+++ src/_HolyDays.js
@@ -1,0 +1,57 @@
+
+export default [
+  {
+     name: 'Naw Ruz',
+     month: 0, //baha
+     day: 1,
+     suspend: true,
+   },
+   {
+     name: 'First Day of Ridván',
+     month: 1, //jalal
+     day: 13,
+     suspend: true,
+   },
+   {
+     name: 'Ninth Day of Ridván',
+     month: 2, //jamal
+     day: 2,
+     suspend: true,
+   },
+   {
+     name: 'Twelfth Day of Ridván',
+     month: 2, //jamal
+     day: 5,
+     suspend: true,
+   },
+   {
+     name: 'Declaration of the Báb',
+     month: 3, //azamat
+     day: 8,
+     suspend: true,
+   },
+   {
+     name: 'Ascension of Bahá’u’lláh',
+     month: 3, //azamat
+     day: 13,
+     suspend: true,
+   },
+   {
+     name: 'Martyrdom of the Báb',
+     month: 5, //rahmat
+     day: 17,
+     suspend: true,
+   },
+   {
+     name: 'Day of the Covenant',
+     month: 13, //qawl
+     day: 4,
+     suspend: false,
+   },
+   {
+     name: 'Ascension of ‘Abdu’l-Bahá',
+     month: 13, //qawl
+     day: 6,
+     suspend: false,
+   },
+]
