+++ src/_LocationMap.js
@@ -1,0 +1,4 @@
+
+export default {
+  Tehran: {latitude: 35.6944, longitude: 51.4215},
+};
