+++ src/examples/example1/img/star.png
