+++ src/examples/example2/index.html
@@ -37,9 +37,7 @@ THE SOFTWARE.
 
 </style>
 
-<script src="extern/stellafane.js"></script>
-<script src="extern/blueyonder.js"></script>
-<script src="badi.js" charset="utf-8"></script>
+<script src="dist/badi.js" charset="utf-8"></script>
 
 <script src="https://code.jquery.com/jquery-1.11.2.min.js"></script>
 <script>
